nbody/requirements.txt
alabaster==0.7.13
appnope==0.1.3
asttokens==2.2.1
Babel==2.12.1
backcall==0.2.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-default-group==1.2.2
cloup==0.13.1
colour==0.1.5
comm==0.1.3
contourpy==1.1.0
cycler==0.11.0
Cython==0.29.35
datetime3==3.8.0
debugpy==1.6.7
decorator==5.1.1
distlib==0.3.6
docopt==0.6.2
docutils==0.20.1
executing==1.2.0
filelock==3.12.2
fonttools==4.40.0
glcontext==2.3.7
idna==3.4
imagesize==1.4.1
ipykernel==6.24.0
ipython==8.14.0
isosurfaces==0.1.0
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.4
manim==0.17.3
ManimPango==0.4.3
mapbox-earcut==1.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
moderngl==5.8.2
moderngl-window==2.4.4
multipledispatch==1.0.0
nest-asyncio==1.5.6
networkx==2.8.8
numpy==1.25.0
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pipreqs==0.4.13
platformdirs==3.8.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pybind11==2.10.4
pycairo==1.24.0
pydub==0.25.1
pyglet==2.0.8
Pygments==2.15.1
pyobjc-core==9.2
pyobjc-framework-Cocoa==9.2
pyopencl==2023.1.1
pyparsing==3.1.0
pyrr==0.10.3
python-dateutil==2.8.2
pytools==2023.1
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
random2==1.0.1
requests==2.31.0
rich==13.4.2
scipy==1.11.1
screeninfo==0.8.1
six==1.16.0
skia-pathops==0.7.4
snowballstemmer==2.2.0
Sphinx==7.0.1
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
srt==3.5.3
stack-data==0.6.2
svgelements==1.9.5
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
tzdata==2023.3
urllib3==2.0.3
virtualenv==20.23.1
watchdog==2.3.1
wcwidth==0.2.6
yarg==0.1.9
